created entities&&services&&controllers for arena&&booking

+++ arena/src/main/java/com/booking/arena/service/address/city/CityServiceImpl.java
@@ -1,12 +1,12 @@
 package com.booking.arena.service.address.city;
 
 
-import com.booking.arena.dto.address.LocationDto;
+import com.booking.arena.dto.address.CityDto;
 import com.booking.arena.entity.address.CityEntity;
 import com.booking.arena.entity.address.CountryEntity;
 import com.booking.arena.exception.ResourceNotFoundException;
-import com.booking.arena.repository.CityRepository;
-import com.booking.arena.repository.CountryRepository;
+import com.booking.arena.repository.address.CityRepository;
+import com.booking.arena.repository.address.CountryRepository;
 import com.booking.arena.utils.ConvertEntityToDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -27,8 +27,8 @@ public class CityServiceImpl implements CityService {
     private final CountryRepository countryRepository;
 
     @Override
-    public Optional<LocationDto> createCity(LocationDto cityDto) {
-        CountryEntity country = countryRepository.findById(cityDto.getId()).orElseThrow(() ->
+    public Optional<CityDto> createCity(CityDto cityDto) {
+        CountryEntity country = countryRepository.findByName(cityDto.getCountryName()).orElseThrow(() ->
                 new ResourceNotFoundException("Not found country with id: " + cityDto.getId()));
         CityEntity city = CityEntity.builder()
                 .name(cityDto.getName())
@@ -38,49 +38,57 @@ public class CityServiceImpl implements CityService {
         city.setUpdatedAt(Instant.now());
         cityRepository.save(city);
         log.info("City created: {}", city.getName());
-        return Optional.of(LocationDto.builder()
+        return Optional.of(CityDto.builder()
                 .name(city.getName())
                 .build());
     }
 
     @Override
-    public Optional<LocationDto> updateCity(Long id, LocationDto cityDto) {
+    public Optional<CityDto> updateCity(Long id, CityDto cityDto) {
         CityEntity city = cityRepository.findById(id).orElseThrow(() ->
                 new ResourceNotFoundException("Not found city with id: " + id));
         city.setName(cityDto.getName());
         city.setUpdatedAt(Instant.now());
         cityRepository.save(city);
         log.info("City updated: {}", city.getName());
-        return Optional.of(LocationDto.builder()
+        return Optional.of(CityDto.builder()
                 .name(city.getName())
                 .build());
     }
 
     @Override
-    public Optional<LocationDto> getCityById(Long id) {
-        return Optional.of(LocationDto.builder()
-                .name(cityRepository.findById(id).orElseThrow(() ->
-                        new ResourceNotFoundException("Not found city with id: " + id)).getName())
+    public Optional<CityDto> getCityById(Long id) {
+        CityEntity city = cityRepository.findById(id).orElseThrow(() ->
+                new ResourceNotFoundException("Not found city with id: " + id));
+        return Optional.of(CityDto.builder()
+                .name(city.getName())
+                .countryName(city.getCountry().getName())
                 .build());
     }
 
     @Override
-    public Optional<LocationDto> getCityByName(String name) {
-        return Optional.of(LocationDto.builder()
-                .name(cityRepository.findByName(name).orElseThrow(() ->
-                        new ResourceNotFoundException("Not found city with id: " + name)).getName())
+    public Optional<CityDto> getCityByName(String name) {
+        CityEntity city = cityRepository.findByName(name).orElseThrow(() ->
+                new ResourceNotFoundException("Not found city with name: " + name));
+        return Optional.of(CityDto.builder()
+                .name(city.getName())
+                .countryName(city.getCountry().getName())
                 .build());
     }
 
     @Override
-    public Optional<List<LocationDto>> getAllCities() {
+    public Optional<List<CityDto>> getAllCities() {
         return Optional.of(cityRepository.findAll().stream()
                 .map(ConvertEntityToDto::cityToDto).collect(Collectors.toList()));
     }
 
     @Override
     public void deleteCityById(Long id) {
-        log.info("City deleted by id: {}", id);
-        cityRepository.deleteById(id);
+        if (cityRepository.existsById(id)) {
+            cityRepository.deleteById(id);
+            log.info("City deleted by id: {}", id);
+        } else {
+            throw new ResourceNotFoundException("Not found city with id: " + id);
+        }
     }
 }
